test: add tests for spherical-polar transition
