+++ src/test/mod.rs
@@ -1,2 +1,1 @@
 mod coordinate_systems;
-mod gnomonic;
